First iteration
This commit represent the two first use cases for craps:

* When Player gets a 7 in the come out roll and wins the game;
* When Player gets a 2 in the come out roll and loses the game;

Right only Game class has some behavior and all test cases for it can be
implemented without any other logic (wondering that I shouldn't
introduce the Player class now).

Introducing the Dice class won't make any difference right now, but it
can be introduced to force cascade tests with Player which would be good
for the sake of the post that contains this code.